Rewrite and modularization
- Removed existing code in MPWindow and moved redundant code to utility classes

Note: this removes all functionality from the tool. it is not in a functioning state right now.
