+++ Scripts/UI/MaterialPalette/ThumbSize.cs
@@ -1,5 +1,6 @@
 ﻿namespace Sabresaurus.SabreCSG.MaterialPalette
 {
+	// TODO: needed?
 	public enum ThumbSize : int
 	{
 		Small = 32,
