+++ Scripts/UI/MaterialPalette/ThumbSize.cs.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
 guid: bf6f70c40ab913846a65573d01c08a10
+timeCreated: 1534043611
+licenseType: Free
 MonoImporter:
-  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
