+++ Scripts/UI/MaterialPalette/Utilities.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f76e1ab53be5faa449043c2079282af3
+folderAsset: yes
+timeCreated: 1536650893
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
