+++ Scripts/UI/MaterialPalette/Utilities/MPGUI.cs.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 9fadbc15793c42e41a895dbaf998f1c3
+guid: dc7f2d4b96459b14fa28d979f0bc3b0d
+timeCreated: 1534043615
+licenseType: Free
 MonoImporter:
-  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
