+++ Scripts/UI/MaterialPalette/Utilities/MPGraphics.cs.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 421f0f53ba2029a4691bbf3be724679c
+guid: f938b16ece1c8ab4cb436269dfc2c954
+timeCreated: 1536650897
+licenseType: Free
 MonoImporter:
-  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
