+++ Scripts/UI/MaterialPalette/Utilities/MPHelper.cs.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 4a532d95248813a4fb6208874f2c7273
+guid: 421f0f53ba2029a4691bbf3be724679c
+timeCreated: 1534043611
+licenseType: Free
 MonoImporter:
-  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
