+++ Scripts/UI/MaterialPalette/Utilities/Styles.cs.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
 guid: 543fb5a5009f98a4a87a46c803582f82
+timeCreated: 1534043611
+licenseType: Free
 MonoImporter:
-  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
