+++ Scripts/UI/MaterialPalette/Window.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 71eed69dba263ee42b57426dea8753b0
+folderAsset: yes
+timeCreated: 1536650942
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
