+++ Scripts/UI/MaterialPalette/Window/MPTagEditorWindow.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d863b551b2a35964ea9102e7bd452561
-timeCreated: 1534043611
+guid: d604c7ffd61bd6d41a06c3ac5372d7be
+timeCreated: 1536651698
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
