+++ Scripts/UI/MaterialPalette/Window/MPWindow.cs.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: dc7f2d4b96459b14fa28d979f0bc3b0d
+guid: 353da3f5cf7588a4b96be79dc5502568
+timeCreated: 1536651698
+licenseType: Free
 MonoImporter:
-  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
