+++ Scripts/UI/PaletteWindow.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d863b551b2a35964ea9102e7bd452561
-timeCreated: 1534043611
+guid: 0e2b8db7a8f26b6439f5eb65963efbda
+timeCreated: 1536647741
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
